moved warehouse destinations to integration package.
