+++ warehouse/integrations/bigquery/bigquery_test.go
@@ -6,6 +6,10 @@ import (
 	"os"
 	"testing"
 
+	"github.com/rudderlabs/rudder-server/warehouse/integrations/testhelper"
+
+	bigquery2 "github.com/rudderlabs/rudder-server/warehouse/integrations/bigquery"
+
 	"cloud.google.com/go/bigquery"
 
 	"github.com/rudderlabs/rudder-server/utils/misc"
@@ -15,9 +19,7 @@ import (
 
 	"github.com/stretchr/testify/require"
 
-	bigquery2 "github.com/rudderlabs/rudder-server/warehouse/bigquery"
 	"github.com/rudderlabs/rudder-server/warehouse/client"
-	"github.com/rudderlabs/rudder-server/warehouse/testhelper"
 	warehouseutils "github.com/rudderlabs/rudder-server/warehouse/utils"
 	"github.com/stretchr/testify/assert"
 )
