+++ warehouse/integrations/clickhouse/testdata/load-copy.csv.gz
