+++ warehouse/integrations/deltalake/deltalake_test.go
@@ -9,14 +9,17 @@ import (
 	"testing"
 	"time"
 
+	"github.com/rudderlabs/rudder-server/warehouse/integrations/testhelper"
+
+	"github.com/rudderlabs/rudder-server/warehouse/integrations/deltalake"
+	"github.com/rudderlabs/rudder-server/warehouse/integrations/deltalake/deltalakeclient"
+
 	"github.com/ory/dockertest/v3"
 	"github.com/rudderlabs/rudder-server/testhelper/destination"
 	"golang.org/x/sync/errgroup"
 
 	"github.com/rudderlabs/rudder-server/config"
 	"github.com/rudderlabs/rudder-server/utils/logger"
-	"github.com/rudderlabs/rudder-server/warehouse/deltalake/deltalakeclient"
-
 	"google.golang.org/grpc"
 
 	proto "github.com/rudderlabs/rudder-server/proto/databricks"
@@ -30,9 +33,6 @@ import (
 	"github.com/rudderlabs/rudder-server/warehouse/client"
 	warehouseutils "github.com/rudderlabs/rudder-server/warehouse/utils"
 	"github.com/stretchr/testify/require"
-
-	"github.com/rudderlabs/rudder-server/warehouse/deltalake"
-	"github.com/rudderlabs/rudder-server/warehouse/testhelper"
 )
 
 func TestIntegrationDeltalake(t *testing.T) {
