+++ warehouse/integrations/docker-compose.test.yml
@@ -233,7 +233,7 @@ services:
       - wh-clickhouse03
       - wh-clickhouse04
     build:
-      context: ../
+      context: ../..
       dockerfile: build/Dockerfile-warehouse-integration
     ports:
       - "8080"
@@ -247,9 +247,9 @@ services:
       - REDSHIFT_INTEGRATION_TEST_SCHEMA
       - SNOWFLAKE_INTEGRATION_TEST_SCHEMA
       - DATABRICKS_INTEGRATION_TEST_SCHEMA
-    entrypoint: sh -c 'go run warehouse/testhelper/generate-workspace-config/generate-workspace-config.go && go run main.go'
+    entrypoint: sh -c 'go run warehouse/integrations/testhelper/generate-workspace-config/generate-workspace-config.go && go run main.go'
     volumes:
-      - ..:/app
+      - ../..:/app
       - /.env
     healthcheck:
       test: wget --no-verbose --tries=1 --spider http://localhost:8080/health || exit 1
