+++ warehouse/integrations/testdata/clickhouse/cluster/clickhouse02/users.xml
