+++ warehouse/integrations/testdata/workspaceConfig/template.json
