+++ warehouse/integrations/testhelper/setup.go
@@ -14,6 +14,12 @@ import (
 	"testing"
 	"time"
 
+	"github.com/rudderlabs/rudder-server/warehouse/integrations/bigquery"
+	"github.com/rudderlabs/rudder-server/warehouse/integrations/deltalake/deltalakeclient"
+	"github.com/rudderlabs/rudder-server/warehouse/integrations/postgres"
+	"github.com/rudderlabs/rudder-server/warehouse/integrations/redshift"
+	"github.com/rudderlabs/rudder-server/warehouse/integrations/snowflake"
+
 	"github.com/rudderlabs/rudder-server/utils/timeutil"
 
 	promCLient "github.com/prometheus/client_model/go"
@@ -22,7 +28,6 @@ import (
 	"github.com/minio/minio-go/v6"
 
 	backendconfig "github.com/rudderlabs/rudder-server/config/backend-config"
-	"github.com/rudderlabs/rudder-server/warehouse/deltalake/deltalakeclient"
 	"github.com/rudderlabs/rudder-server/warehouse/validations"
 
 	"github.com/rudderlabs/rudder-server/utils/httputil"
@@ -31,13 +36,8 @@ import (
 
 	"github.com/cenkalti/backoff"
 
-	"github.com/rudderlabs/rudder-server/config"
-	"github.com/rudderlabs/rudder-server/warehouse/bigquery"
-	"github.com/rudderlabs/rudder-server/warehouse/postgres"
-	"github.com/rudderlabs/rudder-server/warehouse/redshift"
-	"github.com/rudderlabs/rudder-server/warehouse/snowflake"
-
 	_ "github.com/lib/pq"
+	"github.com/rudderlabs/rudder-server/config"
 	"github.com/rudderlabs/rudder-server/warehouse/client"
 	"github.com/stretchr/testify/require"
 )
@@ -70,7 +70,7 @@ const (
 
 const (
 	WorkspaceConfigPath   = "/etc/rudderstack/workspaceConfig.json"
-	WorkspaceTemplatePath = "warehouse/testdata/workspaceConfig/template.json"
+	WorkspaceTemplatePath = "warehouse/integrations/testdata/workspaceConfig/template.json"
 )
 
 type EventsCountMap map[string]int
